[tui] Fix OutputDevice performance on Linux
This commit implements on-demand event-driven write backpressure for
non-blocking stdout on Unix using rustix::event::poll on POLLOUT,
replacing thread timeslice yielding (yield_now()).

Linux rendering for the `cargo run --example tui_apps` ->
`0. 📏 App with no layout` example was 2x faster on macOS than Linux.
And 2x faster on Linux than Windows. This commit fixes this performance
slowdown on Linux makes it even faster than macOS.

1. Performance:
   - Eliminates __sched_yield context-switching thrashing on Linux.
   - Boosts Linux rendering throughput from 106 FPS to 175 FPS (+65%),
     surpassing macOS blocking stdio throughput (163 FPS).
   - Sleep in kernel on PTY wait queue at 0% CPU with sub-microsecond
     wakeups upon buffer availability. This is verified using flamegraph
     analysis.

2. Modular Architecture in core/terminal_io/:
   - Extracted BackpressureStdout into backpressure_stdout/ submodule
     (backpressure_stdout_struct.rs, impl_unix.rs, impl_win.rs).
   - Extracted TerminalModeController and FullScreenTuiModeGuard into
     terminal_mode_controller.rs.
   - Renamed terminal_io_type_aliases.rs to types.rs.
   - Maintained 100% backward-compatible flat public barrel exports.

3. Dependencies & Tooling:
   - Enabled "event" feature on rustix in tui/Cargo.toml.
   - Added self-upgrade support scripts and metadata.

Task: prepare-v0.8.0-meta-task.md,
      fix-yield_now-slowdown-on-linux.md
