--- .vscode/bookmarks.json
@@ -1,24 +1,0 @@
-{
-	"files": [
-		{
-			"path": "tui/src/tui/editor/editor_engine/validate_scroll_on_resize.rs",
-			"bookmarks": [
-				{
-					"line": 85,
-					"column": 0,
-					"label": "this is where code smell exists between old and new canvas & viewport world - should use viewport bounds check"
-				}
-			]
-		},
-		{
-			"path": "task/modernize-editor-using-new-units.md",
-			"bookmarks": [
-				{
-					"line": 403,
-					"column": 0,
-					"label": "We stopped work here to jump off into viewport_bounds_check.md"
-				}
-			]
-		}
-	]
-}
