+++ .vscode/settings.json
@@ -297,6 +297,7 @@
     "rapidblink",
     "rclone",
     "Rearchitect",
+    "redirections",
     "reedline",
     "Renderable",
     "renderop",
