+++ tui/Cargo.toml
@@ -46,8 +46,11 @@ chrono = "0.4.42"
 crossterm = { version = "0.29.0", features = ["event-stream"] }
 futures-util = "0.3.31"
 
-# Terminal raw mode support with safe Rust APIs.
-rustix = { version = "1.1.2", features = ["termios", "std"] }
+# Safe Rust POSIX APIs:
+# - termios: raw mode terminal state manipulation.
+# - std: std::io and standard library type conversions.
+# - event: event-driven POLLOUT backpressure via poll(2).
+rustix = { version = "1.1.2", features = ["termios", "std", "event"] }
 
 # mio for cross-platform async I/O (uses kqueue on macOS, epoll on Linux).
 # Used by DirectToAnsi backend to poll stdin and SIGWINCH together.
