+++ tui/src/core/terminal_io/types.rs
